Testing out data augmentation and atan2 implementation for new flow viz.

+++ src/util.py
@@ -92,30 +92,49 @@ def draw_hsv(flow):
     return bgr
 
 
-def draw_flow(img, flow, step=16):
-    img = cv2.cvtColor(cv2.imread(img), cv2.COLOR_BGR2GRAY)
-    h, w = img.shape[:2]
-    y, x = np.mgrid[step/2:h:step, step/2:w:step].reshape(2, -1).astype(int)
-    fx, fy = flow[y, x].T*5
-    lines = np.vstack([x, y, x+fx, y+fy]).T.reshape(-1, 2, 2)
-    lines = np.int32(lines + 0.5)
-    vis = cv2.cvtColor(img, cv2.COLOR_GRAY2BGR)
-    cv2.polylines(vis, lines, 0, (0, 255, 0))
-    for (x1, y1), (x2, y2) in lines:
-        cv2.circle(vis, (x1, y1), 1, (0, 255, 0), -1)
+def atan2_ocv(y, x):
+    # constants
+    DBL_EPSILON = 2.2204460492503131e-16
+    atan2_p1 = 0.9997878412794807 * (180 / np.pi)
+    atan2_p3 = -0.3258083974640975 * (180 / np.pi)
+    atan2_p5 = 0.1555786518463281 * (180 / np.pi)
+    atan2_p7 = -0.04432655554792128 * (180 / np.pi)
 
-    return vis
+    ax, ay = tf.abs(x), tf.abs(y)
+    c = tf.select(tf.greater_equal(ax, ay), tf.div(ay, ax + DBL_EPSILON),
+                  tf.div(ax, ay + DBL_EPSILON))
+    c2 = tf.square(c)
+    angle = (((atan2_p7 * c2 + atan2_p5) * c2 + atan2_p3) * c2 + atan2_p1) * c
+    angle = tf.select(tf.greater_equal(ax, ay), angle, 90.0 - angle)
+    angle = tf.select(tf.less(x, 0.0), 180.0 - angle, angle)
+    angle = tf.select(tf.less(y, 0.0), 360.0 - angle, angle)
+    return angle
 
 
-def warp_flow(img, flow):
-    img = cv2.imread(img)
-    h, w = flow.shape[:2]
-    flow = -flow
-    flow[:, :, 0] += np.arange(w)
-    flow[:, :, 1] += np.arange(h)[:, np.newaxis]
-    res = cv2.remap(img, flow, None, cv2.INTER_CUBIC)
+def normalize(tensor, a=0, b=1):
+    return tf.div(tf.mul(tf.sub(tensor, tf.reduce_min(tensor)), b - a),
+                  tf.sub(tf.reduce_max(tensor), tf.reduce_min(tensor)))
 
-    return res
+
+def cart_to_polar_ocv(x, y, angle_in_degrees=False):
+    v = tf.sqrt(tf.add(tf.square(x), tf.square(y)))
+    ang = atan2_ocv(y, x)
+    scale = 1 if angle_in_degrees else np.pi / 180
+    return v, tf.mul(ang, scale)
+
+
+def draw_hsv_ocv(flow):
+    fx, fy = flow[:, :, :, 0], flow[:, :, :, 1]
+    v, ang = cart_to_polar_ocv(fx, fy)
+
+    h = normalize(tf.mul(ang, 180 / np.pi))
+    s = tf.ones_like(h)
+    v = normalize(v)
+
+    hsv = tf.pack([h, s, v], 3)
+    rgb = tf.image.hsv_to_rgb(hsv) * 255
+
+    return tf.cast(rgb, tf.uint8)
 
 
 def check_snapshots(root_folder='', train=True):
@@ -145,10 +164,16 @@ def check_snapshots(root_folder='', train=True):
                           ' scratch'
                     resume = False
                     for file in os.listdir(snapshots_folder):
+                        if file == '.gitignore':
+                            continue
                         os.remove(snapshots_folder + file)
                     for file in os.listdir(logs_folder + 'train/'):
+                        if file == '.gitignore':
+                            continue
                         os.remove(logs_folder + 'train/' + file)
                     for file in os.listdir(logs_folder + 'val/'):
+                        if file == '.gitignore':
+                            continue
                         os.remove(logs_folder + 'val/' + file)
         else:
             print "No snapshots found, training from scratch"
